Moved types.ts which was previously missed
