+++ app/src/lib/backend/aiClient/butler.ts
@@ -1,6 +1,6 @@
 import { MessageRole, type AIClient, type PromptMessage } from '$lib/backend/aiClient';
 import type { CloudClient } from '$lib/backend/cloud';
-import type { ModelKind } from '../types';
+import type { ModelKind } from './types';
 
 export class ButlerAIClient implements AIClient {
 	constructor(
