+++ app/src/lib/components/AISettings.svelte
@@ -5,9 +5,9 @@
 	import Spacer from './Spacer.svelte';
 	import TextBox from './TextBox.svelte';
 	import WelcomeSigninAction from './WelcomeSigninAction.svelte';
+	import { OpenAIModelName, AnthropicModelName, ModelKind } from '$lib/backend/aiClient/types';
 	import { AIService, GitAIConfigKey, KeyOption } from '$lib/backend/aiService';
 	import { GitConfigService } from '$lib/backend/gitConfigService';
-	import { OpenAIModelName, AnthropicModelName, ModelKind } from '$lib/backend/types';
 	import RadioButton from '$lib/components/RadioButton.svelte';
 	import SectionCard from '$lib/components/SectionCard.svelte';
 	import { UserService } from '$lib/stores/user';
